requirements.txt
torch
torchvision
opencv-python
Pillow
numpy
insightface
onnxruntime
onnxruntime-gpu
pyinstaller


